修改加载的bug

+++ Assets/Scripts/UI/LoadingScene.cs
@@ -33,7 +33,7 @@ public class LoadingScene : MonoBehaviour
         if (isStart)
         {
             progressBar.value = ao.progress;
-            progressShow.text = ao.progress * 100 + "%";
+            progressShow.text = (int) ao.progress * 100 + "%";
             if (ao.isDone)
             {
                 progressBar.gameObject.SetActive(false);
